+++ frontend/svelte/src/lib/components/sns-launchpad/ProposalCard.svelte
