+++ frontend/svelte/src/tests/routes/Launchpad.spec.ts
@@ -3,12 +3,12 @@
  */
 
 import { render } from "@testing-library/svelte";
-import LaunchPad from "../../routes/Launchpad.svelte";
+import Launchpad from "../../routes/Launchpad.svelte";
 import en from "../mocks/i18n.mock";
 
 describe("Launchpad", () => {
   it("should render titles", () => {
-    const { getByText } = render(LaunchPad);
+    const { getByText } = render(Launchpad);
 
     expect(getByText(en.sns_launchpad.projects)).toBeInTheDocument();
     expect(getByText(en.sns_launchpad.proposals)).toBeInTheDocument();
